Move unique password policy tasks and signals into policies.unique_password app
